feat/move sessions helper to api
